boss, baby elf, n enemy
