+++ goblin.py
@@ -1,6 +1,7 @@
 import random
+from enemy import Enemy
 
-class Goblin:
+class Goblin(Enemy):
     """
     This is our goblin blueprint 
     
@@ -9,21 +10,7 @@ class Goblin:
         health: The current health value 
         attack_power: How much health will be drained from opponent if hit
     """
-    def __init__(self, name):
-        self.name = name
-        self.health = 100
-        self.attack_power = random.randint(5, 15)
+    def __init__(self, name, color):
+        super().__init__(name)
+        self.color = color
 
-    def attack(self):
-        return random.randint(1, self.attack_power)
-
-    def take_damage(self, damage):
-        self.health -= damage
-        if self.health < 0:
-            new_health = 0
-        else:
-            new_health = self.health
-        print(f"{self.name} takes {damage} damage. Health is now {new_health}.")
-
-    def is_alive(self):
-        return self.health > 0
